requirements.txt
playwright
google-generativeai
chromadb
markdown
typer
chromadb
gtts
pydub
python-vlc
python-Levenshtein
asyncio
streamlit
#vlc media player for audio playback